+++ src/FilterRow.tsx
@@ -4,12 +4,12 @@ import classNames from 'classnames';
 import { CalculatedColumn, Filters } from './common/types';
 import { DataGridProps } from './DataGrid';
 
-type SharedDataGridProps<R, K extends keyof R> = Pick<DataGridProps<R, K>,
+type SharedDataGridProps<R> = Pick<DataGridProps<R, never>,
 | 'filters'
 | 'onFiltersChange'
 >;
 
-export interface FilterRowProps<R, K extends keyof R> extends SharedDataGridProps<R, K> {
+export interface FilterRowProps<R> extends SharedDataGridProps<R> {
   height: number;
   width: number;
   lastFrozenColumnIndex: number;
@@ -17,7 +17,7 @@ export interface FilterRowProps<R, K extends keyof R> extends SharedDataGridProp
   scrollLeft: number | undefined;
 }
 
-export default function FilterRow<R, K extends keyof R>({
+export default function FilterRow<R>({
   height,
   width,
   columns,
@@ -25,9 +25,9 @@ export default function FilterRow<R, K extends keyof R>({
   scrollLeft,
   filters,
   onFiltersChange
-}: FilterRowProps<R, K>) {
-  function onChange(key: keyof R, value: unknown) {
-    const newFilters: Filters<R> = { ...filters };
+}: FilterRowProps<R>) {
+  function onChange(key: string, value: unknown) {
+    const newFilters: Filters = { ...filters };
     newFilters[key] = value;
     onFiltersChange?.(newFilters);
   }
@@ -55,7 +55,7 @@ export default function FilterRow<R, K extends keyof R>({
 
         return (
           <div
-            key={key as string}
+            key={key}
             style={style}
             className={className}
           >