+++ CHANGELOG.md
@@ -73,6 +73,7 @@
   - ⚠️ Started publishing ES2020/ESM modules instead of ES5/CommonJS modules.
     - Using [`@babel/preset-env`](https://www.npmjs.com/package/@babel/preset-env) with [`core-js`](https://www.npmjs.com/package/core-js) is recommended to enable compatibility for the browsers your project aims to support.
   - Only visible headers cells are now rendered. [#1837](https://github.com/adazzle/react-data-grid/pull/1837)
+  - ⚠️ the `rowKey` prop is now required for row selection.
   - ⚠️ `column.cellClass` does not affect header cells anymore.
 
 ## `master` to `alpha`
@@ -88,7 +89,6 @@
     - ⚠️ Row selection has been reimplemented.
     - A new `SelectColumn` is now available to import and add a row selection column.
     - Check [#1762](https://github.com/adazzle/react-data-grid/pull/1762) on how to migrate.
-  - `renderBatchSize` prop [#1779](https://github.com/adazzle/react-data-grid/pull/1779)
 - **Removed:**
   - **Packages:**
     - `react-data-grid-examples`